+++ OpenSimplex2S/test/bitmap.h
@@ -1,4 +1,4 @@
 #define WIDTH 512
 #define HEIGHT 512
 
-void save_bitmap(char *filename, float **vals);
+void save_bitmap(char *filename, double **vals);
